+++ Gw2Sharp/WebApi/V2/Clients/Continents/ContinentsFloorsRegionsMapsIdClient.cs
@@ -1,4 +1,5 @@
 using System;
+using Gw2Sharp.WebApi.V2.Models;
 
 namespace Gw2Sharp.WebApi.V2.Clients
 {
@@ -6,7 +7,11 @@ namespace Gw2Sharp.WebApi.V2.Clients
     /// A client of the Guild Wars 2 API v2 continents floors regions maps id endpoint.
     /// </summary>
     [EndpointPath("continents/:continent_id/floors/:floor_id/regions/:region_id/maps/:map_id")]
-    public class ContinentsFloorsRegionsMapsIdClient : BaseClient, IContinentsFloorsRegionsMapsIdClient
+    [EndpointPathSegment("continent_id", 0)]
+    [EndpointPathSegment("floor_id", 1)]
+    [EndpointPathSegment("region_id", 2)]
+    [EndpointPathSegment("map_id", 3)]
+    public class ContinentsFloorsRegionsMapsIdClient : BaseEndpointBlobClient<ContinentFloorRegionMap>, IContinentsFloorsRegionsMapsIdClient
     {
         private readonly int continentId;
         private readonly int floorId;
@@ -27,7 +32,7 @@ namespace Gw2Sharp.WebApi.V2.Clients
         /// <param name="mapId">The map id.</param>
         /// <exception cref="ArgumentNullException"><paramref name="connection"/> or <paramref name="gw2Client"/> is <c>null</c>.</exception>
         protected internal ContinentsFloorsRegionsMapsIdClient(IConnection connection, IGw2Client gw2Client, int continentId, int floorId, int regionId, int mapId) :
-            base(connection, gw2Client)
+            base(connection, gw2Client, continentId.ToString(), floorId.ToString(), regionId.ToString(), mapId.ToString())
         {
             this.continentId = continentId;
             this.floorId = floorId;